+++ src/layouts/dashboard/Header/HeaderMenuOutlet/style.js
@@ -1,0 +1,6 @@
+// Material UI
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({}));
+
+export default useStyles;