+++ src/store/layout/reducers.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import dashboard from './dashboard/reducers';
+
+const reducers = combineReducers({
+  dashboard,
+});
+
+export default reducers;