+++ src/store/modules/user/actions.js
@@ -1,0 +1,11 @@
+// Types
+import * as types from './types';
+
+export const actionModuleUserStatus = (status) => (dispatch) => {
+  return dispatch({
+    type: types.STATUS,
+    payload: {
+      status,
+    },
+  });
+};